Collects and Generates documentation of a package in a tibble format.

This function crawls and aggregates roxygen documentation into a tibble format. To work properly, elements must be separated with the named fields at title, at description, at ...), each at will be used as column name. The column name will also have 80 character to show the margin limit of each chunk of documentation.

collect_roxygen(folder_r = "R")

Arguments

folder_r

A character string identifying the folder to index. If not specified, 'R/' is the default.

Value

A tibble where each line represents a function described in a package, and each column is documentation field. Most common fields (title, description, details, param, see also, return and examples are placed ahead).

Examples

{

library(tidyr)
try({tibble(collect_roxygen(tempfile()))}, silent = FALSE)

}
#> Your files contained in your R environment are currently being indexed.
#>     Please wait...
#> Error in tools::file_path_as_absolute(folder) : 
#>   file 'C:\Users\guill\AppData\Local\Temp\RtmpQdDxzc\file94b012ac2911' does not exist