Create summary tibble of a (possibly grouped) numerical-type column

This function creates datatable of the values of a column separating valid, non-valid and missing values. Missing values can be given as input to non-valid and valid values separately, or grouped by another column. The output can be editable (using plotly library) or static (using ggplot2 library). The R-code is also editable for coding recycling purpose. The user can download the datatable in csv format.

summary_numerical(
  tbl = "iris",
  col = "col",
  filter = "c()",
  negate = FALSE,
  missing_values = "c()",
  out = "DT-cat",
  group_by = NULL
)

Arguments

tbl

A character string or tibble

col

A character string of a column of interest

filter

A character string to subset the rows, applying the expressions to the column values to determine which rows should be retained. It can be applied to both grouped and ungrouped data.

negate

If TRUE, return non-matching elements.

missing_values

Vector listing values to exclude from valid values. Those values will not be excluded from counting, but will be separated from valid values.

out

parameter that specifies the output expected: can be either 'DT', 'DT-code' and 'DT-cat'. DT renders a datatable using DT library, code gives the code in a string (usable directly with eval/parse functions) and cat provides indented code in the console.

group_by

A character string of one column in the tbl that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

Value

A datatable (editable) object or a R script in a character string to create it.

See also

Examples

{

##### Example 1 -------------------------------------------------------------
# cat output generated as a template when no argument provided
summary_numerical()

##### Example 2 -------------------------------------------------------------
# summary table of Petal.Length
summary_numerical(tbl = iris, col = "Petal.Length", out = "DT")

}
#> iris %>% 
#>   filter(!(col %in% c())) %>% 
#>   group_by('') %>% 
#>   select(col) %>% 
#>   rename(group = 1) %>% 
#>   mutate(group = ifelse(group == '', ' ', group)) %>% 
#>   summarise( 
#>     nbr.val = sum(!is.na(col)), 
#>     nbr.na  = sum(is.na(col)), 
#>     min = min(col, na.rm = TRUE), 
#>     max = max(col, na.rm = TRUE), 
#>     range = max - min, 
#>     median = median(col, na.rm = TRUE), 
#>     mean = mean(col, na.rm = TRUE), 
#>     std.dev = sd(col, na.rm = TRUE)) %>% 
#>   tidyr::pivot_longer(!group) %>% 
#>   tidyr::pivot_wider(names_from = group, 
#>   names_glue = '{group}<br>(all answers)') %>%
#>  full_join(     
#> iris %>% 
#>   filter(!(col %in% c())) %>% 
#>   filter(!(col %in% c()| is.na(col))) %>% 
#>   group_by('') %>% 
#>   select(col) %>% 
#>   rename(group = 1) %>% 
#>   mutate(group = ifelse(group == '', ' ', group)) %>% 
#>   summarise( 
#>     nbr.val = sum(!is.na(col)), 
#>     nbr.na  = sum(is.na(col)), 
#>     min = min(col, na.rm = TRUE), 
#>     max = max(col, na.rm = TRUE), 
#>     range = max - min, 
#>     median = median(col, na.rm = TRUE), 
#>     mean = mean(col, na.rm = TRUE), 
#>     std.dev = sd(col, na.rm = TRUE)) %>% 
#>   tidyr::pivot_longer(!group) %>% 
#>   tidyr::pivot_wider(names_from = group, 
#>   names_glue = '{group}<br>(only valid answers)')) %>%
#>   select(' ' = name, order(colnames(.))) %>% 
#>   mutate_at(.vars = -1, ~ round(., 2)) %>% 
#>   DT::datatable( 
#>     class = 'cell-border stripe', rownames = TRUE,
#>     editable = FALSE, extensions = 'Buttons',
#>     options = list(scrollX=TRUE,dom='Brtip',buttons='csv'),escape=FALSE)
#> Adding missing grouping variables: `""`
#> Adding missing grouping variables: `""`
#> Joining with `by = join_by(name)`